Calculate prediction intervals as proportions
Simply using raw volumes and subtracting them from the prediction
to find the anomaly "threshold" has some flaws:

 * we are only combining all timestamps into 12 bins, which
   means we group together times where the signal values may be
   wildly different (e.g. consider traffic volume at 2pm vs 2am,
   which is likely to end up in the same "bin").
 * we use the median of all prediction errors in the bin as the
   basis for our threshold, we can end up with a threshold that
   is way too low for times when the signal is at its lowest,
   and probably also too high for the times when the signal is
   at its peak.
 * we can't realistically calculate individual thresholds for
   each separate time slot within a season, as we'd need a TON
   of history to get enough data points to make a good estimate.

So, I'm trying out an approach where we instead look at the
initial prediction errors as a proportion of the actual value.
This helps alleviate (somewhat) the issues from mixing very
different time periods in the same bin and since we still use
the median multiplied by 1.428 (as a stddev approximation),
and then multiplied further by 3 to try and emulate a 99%
confidence interval, we're still going to have very generous
thresholds.

They just won't be quite as generous in the past in a lot of
cases, so I'll need to do a bit more testing.

I also need to update the actual libchocolatine code to
calculate thresholds using the new proportionate pred_intervals.